+++ src/main/java/dev/enjarai/restartdetector/ModMath.java
@@ -1,0 +1,6 @@
+package dev.enjarai.restartdetector;
+
+public class ModMath {
+    public static final float TO_RAD = (float) (Math.PI / 180);
+    public static final float TO_DEG = 1 / TO_RAD;
+}